Fixing all loops in all dxf files
Almost all DXF files had some loops in it (lines or dots that were not connected to anything else). That was a problem, as most laser cutter will act weirdly because of that.
Fixing this allow laser cutter to do the cut in one smooth line, and not try to leave some very close gaps because some dots were not connected.

Files were imported then exported using Fusion 360. It seems that they have been vastly re-written, but as the DXF format has been invented by Autodesk, it should be ok and readable by other CAD softwares.

You can see an example of the loops in one file here : https://imgur.com/zZA2Odz
This is checked using an addin called Sketch Checker in Fusion 360.